Weapon 2 finished
The second weapon, a beam weapon, is meant to be stronger against goliaths by piercing their shields, and weaker against scouts. It has a relatively low fire rate, but does tons of damage when both beams hit a target at the same time. This is opposed to weapon 1, which is meant to be able to fight large amounts of enemies, instead of very high health enemies.