Screenshake and minor soundfix.
Everytime the player gets hit by a enemy bullet, the screen shakes a bit so the player has an easier time recognizing they've been hit. You can also summon the function at any time for any event should you wish to do so in the code. The magnitude of the shakes is customizable when calling the function.